+++ src/main/java/nikdevs/onlinestore/service/interfaces/BrandService.java
@@ -1,13 +1,13 @@
 package nikdevs.onlinestore.service.interfaces;
 
-import nikdevs.onlinestore.service.model.BrandRepr;
+import nikdevs.onlinestore.service.repr.BrandRepr;
 
 import java.util.List;
 
 public interface BrandService {
 
     List<BrandRepr> findAll();
-    BrandRepr findById(int id);
+    BrandRepr findById(Long id);
     void save(BrandRepr brand);
-    void remove(int id);
+    void remove(Long id);
 }
